+++ CHANGELOG.md
@@ -15,5 +15,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Create draft Professions in the database, rather than using session storage when adding a new profession
 - Fix bug where submitting invalid top level details on a new Profession caused the page to hang
+- Add functionality to "change" links on Check your answers when creating a new Profession
 
 [unreleased]: https://github.com/UKGovernmentBEIS/regulated-professions-register/compare/release...HEAD